move AdminAPI to tools module, created README.md file and refactored the code to use the properties of CommandClient instead of create those instances again.
